small fixes

+++ src/main/java/com/edolk/natives/functions/Random.java
@@ -22,7 +22,7 @@ public class Random implements Callable {
         Object arg0 = arguments.get(0);
         Object arg1 = arguments.get(1);
         if (arg0 instanceof Double min && arg1 instanceof Double max) {
-            return new java.util.Random().nextInt(((int)Math.floor(max) - (int)Math.floor(min)) + 1) + (int)Math.floor(min);
+            return (double)new java.util.Random().nextInt(((int)Math.floor(max) - (int)Math.floor(min)) + 1) + (int)Math.floor(min);
         }
         return null;
     }
